Added Man Pages. Fixed output day rows. Fixed years.in : if 365 and leap year then makes it 366 for last day of last year.

+++ R/C_swYears.R
@@ -4,7 +4,7 @@
 ###############################################################################
 
 print("swYears")
-swYears<-setClass("swYears",representation(StartYear="integer",EndYear="integer",FDOFY="integer",EDOEY="integer",isNorth="logical"),prototype=prototype(StartYear=as.integer(1982),EndYear=as.integer(1990),FDOFY=as.integer(1),EDOEY=as.integer(365),isNorth=TRUE))
+swYears<-setClass("swYears",representation(StartYear="integer",EndYear="integer",FDOFY="integer",EDOEY="integer",isNorth="logical"),prototype=prototype(StartYear=as.integer(1979),EndYear=as.integer(2010),FDOFY=as.integer(1),EDOEY=as.integer(365),isNorth=TRUE))
 swYears_validity<-function(object){
 	if(length(object@StartYear)!=1 | is.na(object@StartYear) | object@StartYear<0)
 		return("@StartYear:Needstobeintegerlength1.Non-negativeandnotNA")	
@@ -19,7 +19,7 @@ swYears_validity<-function(object){
 	TRUE
 }
 setValidity("swYears",swYears_validity)
-setMethod("initialize","swYears",function(.Object,StartYear=as.integer(1982),EndYear=as.integer(1986),FDOFY=as.integer(1),EDOEY=as.integer(365),isNorth=TRUE){
+setMethod("initialize","swYears",function(.Object,StartYear=as.integer(1979),EndYear=as.integer(2010),FDOFY=as.integer(1),EDOEY=as.integer(365),isNorth=TRUE){
 			.Object@StartYear=StartYear
 			.Object@EndYear=EndYear
 			.Object@FDOFY=FDOFY
